move package table group
