+++ src/test/java/kitchenpos/application/TableGroupServiceTest.java
@@ -8,15 +8,16 @@ import java.util.Collections;
 import java.util.List;
 import java.util.stream.Stream;
 import kitchenpos.order.application.dto.OrderTableRequest;
-import kitchenpos.application.dto.TableGroupRequest;
-import kitchenpos.application.dto.TableGroupResponse;
+import kitchenpos.tableGroup.application.TableGroupService;
+import kitchenpos.tableGroup.application.dto.TableGroupRequest;
+import kitchenpos.tableGroup.application.dto.TableGroupResponse;
 import kitchenpos.order.domain.repository.OrderDao;
 import kitchenpos.order.domain.repository.OrderTableDao;
-import kitchenpos.dao.TableGroupDao;
+import kitchenpos.tableGroup.repository.TableGroupDao;
 import kitchenpos.order.domain.Order;
 import kitchenpos.order.domain.OrderStatus;
 import kitchenpos.order.domain.OrderTable;
-import kitchenpos.domain.TableGroup;
+import kitchenpos.tableGroup.domain.TableGroup;
 import kitchenpos.support.ServiceTest;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
