+++ src/test/java/kitchenpos/application/TableServiceTest.java
@@ -10,11 +10,11 @@ import kitchenpos.order.application.dto.OrderTableRequest.NumberOfGuest;
 import kitchenpos.order.application.dto.OrderTableResponse;
 import kitchenpos.order.domain.repository.OrderDao;
 import kitchenpos.order.domain.repository.OrderTableDao;
-import kitchenpos.dao.TableGroupDao;
+import kitchenpos.tableGroup.repository.TableGroupDao;
 import kitchenpos.order.domain.Order;
 import kitchenpos.order.domain.OrderStatus;
 import kitchenpos.order.domain.OrderTable;
-import kitchenpos.domain.TableGroup;
+import kitchenpos.tableGroup.domain.TableGroup;
 import kitchenpos.support.ServiceTest;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
